pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-client-generator"
version = "0.1.0"  # Remember to update this when you release new versions
description = "Automatically generate Python API clients from OpenAPI 3.1.0 specifications"
readme = "README.md"
authors = [{ name = "HonestSoul", email = "brij.pydata@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["api", "client", "generator", "openapi", "swagger"]
dependencies = [
    "requests>=2.26.0",
    "aiohttp>=3.7.4",
    "jinja2>=3.0.1",
    "pyyaml>=6.0",
    "jsonschema>=4.4.0",
    "python-jose>=3.3.0",
]
requires-python = ">=3.7"

[project.optional-dependencies]
dev = ["pytest", "black", "isort"]

[project.urls]
Homepage = "https://github.com/honestsoul/python-api-client-generator"
"Bug Tracker" = "https://github.com/honestsoul/python-api-client-generator/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["api_client_generator*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
api_client_generator = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.isort]
profile = "black"